Module aws_sdk_lookoutvision::error

Modules

access_denied_exception

See AccessDeniedException

conflict_exception

See ConflictException

internal_server_exception

See InternalServerException

resource_not_found_exception

See ResourceNotFoundException

service_quota_exceeded_exception

See ServiceQuotaExceededException

throttling_exception

See ThrottlingException

validation_exception

See ValidationException

Structs

AccessDeniedException

You are not authorized to perform the action.

ConflictException

The update or deletion of a resource caused an inconsistent state.

CreateDatasetError

CreateModelError

CreateProjectError

DeleteDatasetError

DeleteModelError

DeleteProjectError

DescribeDatasetError

DescribeModelError

DescribeProjectError

DetectAnomaliesError

InternalServerException

Amazon Lookout for Vision experienced a service issue. Try your call again.

ListDatasetEntriesError

ListModelsError

ListProjectsError

ListTagsForResourceError

ResourceNotFoundException

The resource could not be found.

ServiceQuotaExceededException

A service quota was exceeded the allowed limit. For more information, see

StartModelError

StopModelError

TagResourceError

ThrottlingException

Amazon Lookout for Vision is temporarily unable to process the request. Try your call again.

UntagResourceError

UpdateDatasetEntriesError

ValidationException

An input validation error occured. For example, invalid characters in a project name,

Enums

CreateDatasetErrorKind

CreateModelErrorKind

CreateProjectErrorKind

DeleteDatasetErrorKind

DeleteModelErrorKind

DeleteProjectErrorKind

DescribeDatasetErrorKind

DescribeModelErrorKind

DescribeProjectErrorKind

DetectAnomaliesErrorKind

ListDatasetEntriesErrorKind

ListModelsErrorKind

ListProjectsErrorKind

ListTagsForResourceErrorKind

StartModelErrorKind

StopModelErrorKind

TagResourceErrorKind

UntagResourceErrorKind

UpdateDatasetEntriesErrorKind