Module aws_sdk_lookoutvision::model

Modules

dataset_description

See DatasetDescription

dataset_ground_truth_manifest

See DatasetGroundTruthManifest

dataset_image_stats

See DatasetImageStats

dataset_metadata

See DatasetMetadata

dataset_source

See DatasetSource

detect_anomaly_result

See DetectAnomalyResult

image_source

See ImageSource

input_s3_object

See InputS3Object

model_description

See ModelDescription

model_metadata

See ModelMetadata

model_performance

See ModelPerformance

output_config

See OutputConfig

output_s3_object

See OutputS3Object

project_description

See ProjectDescription

project_metadata

See ProjectMetadata

s3_location

See S3Location

tag

See Tag

Structs

DatasetDescription

The description for a dataset. For more information, see DescribeDataset.

DatasetGroundTruthManifest

Location information about a manifest file. You can use a manifest file to

DatasetImageStats

Statistics about the images in a dataset.

DatasetMetadata

Sumary information for an Amazon Lookout for Vision dataset.

DatasetSource

Information about the location of a manifest file that Amazon Lookout for Vision uses to to create a dataset.

DetectAnomalyResult

The prediction results from a call to DetectAnomalies.

ImageSource

The source for an image.

InputS3Object

Amazon S3 Location information for an input manifest file.

ModelDescription

Describes an Amazon Lookout for Vision model.

ModelMetadata

Describes an Amazon Lookout for Vision model.

ModelPerformance

Information about the evaluation performance of a trained model.

OutputConfig

The S3 location where Amazon Lookout for Vision saves model training files.

OutputS3Object

The S3 location where Amazon Lookout for Vision saves training output.

ProjectDescription

Describe an Amazon Lookout for Vision project. For more information, see DescribeProject.

ProjectMetadata

Metadata about an Amazon Lookout for Vision project.

S3Location

Information about the location training output.

Tag

A key and value pair that is attached to the specified Amazon Lookout for Vision model.

Enums

DatasetStatus

ModelHostingStatus

ModelStatus

ResourceType