Module aws_sdk_lookoutvision::operation

Structs

CreateDataset

Creates a new dataset in an Amazon Lookout for Vision project. CreateDataset can create a

CreateModel

Creates a new version of a model within an an Amazon Lookout for Vision project.

CreateProject

Creates an empty Amazon Lookout for Vision project. After you create the project, add a dataset by calling

DeleteDataset

Deletes an existing Amazon Lookout for Vision dataset.

DeleteModel

Deletes an Amazon Lookout for Vision model. You can't delete a running model. To stop a running model,

DeleteProject

Deletes an Amazon Lookout for Vision project.

DescribeDataset

Describe an Amazon Lookout for Vision dataset.

DescribeModel

Describes a version of an Amazon Lookout for Vision model.

DescribeProject

Describes an Amazon Lookout for Vision project.

DetectAnomalies

Detects anomalies in an image that you supply.

ListDatasetEntries

Lists the JSON Lines within a dataset. An Amazon Lookout for Vision JSON Line contains the anomaly

ListModels

Lists the versions of a model in an Amazon Lookout for Vision project.

ListProjects

Lists the Amazon Lookout for Vision projects in your AWS account.

ListTagsForResource

Returns a list of tags attached to the specified Amazon Lookout for Vision model.

StartModel

Starts the running of the version of an Amazon Lookout for Vision model. Starting a model takes a while

StopModel

Stops the hosting of a running model. The operation might take a while to complete. To

TagResource

Adds one or more key-value tags to an Amazon Lookout for Vision model.

UntagResource

Removes one or more tags from an Amazon Lookout for Vision model. For more information, see

UpdateDatasetEntries

Adds one or more JSON Line entries to a dataset. A JSON Line includes information about an image